requirements.txt
Django==1.5.4
SimpleAES==1.0
South==0.8.2
Werkzeug==0.9.4
argparse==1.2.1
coverage==3.6
django-coverage==1.2.4
django-crispy-forms==1.4.0
django-debug-toolbar==0.9.4
django-extensions==1.2.2
django-nose==1.2
django-password==3.0
django-passwords==0.2.0
django-simple-aes-field==0.1.3
factory-boy==2.2.1
nose==1.3.0
pycrypto==2.6
six==1.4.1
wsgiref==0.1.2
